% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chili.R
\name{chili}
\alias{chili}
\title{Chili is the equation for force}
\usage{
chili(mass, acceleration)
}
\arguments{
\item{mass}{This is the mass}

\item{acceleration}{This is the acceleration}
}
\description{
Chili is the equation for force
}
